move hidden gallery button to the main gallery
